int id in repos

+++ src/main/java/kr/di/uoa/gr/jedaiwebapp/models/DatasetRepository.java
@@ -2,7 +2,7 @@ package kr.di.uoa.gr.jedaiwebapp.models;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface DatasetRepository extends CrudRepository<Dataset, Double>{
+public interface DatasetRepository extends CrudRepository<Dataset, Integer>{
 		
 	Dataset findById(int id);
