+++ src/main/java/kr/di/uoa/gr/jedaiwebapp/models/WorkflowConfigurationRepository.java
@@ -3,7 +3,7 @@ package kr.di.uoa.gr.jedaiwebapp.models;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface WorkflowConfigurationRepository extends CrudRepository<WorkflowConfiguration, Double>{
+public interface WorkflowConfigurationRepository extends CrudRepository<WorkflowConfiguration, Integer>{
 	
 	WorkflowConfiguration findById(int id);
