+++ src/main/java/kr/di/uoa/gr/jedaiwebapp/models/WorkflowResultsRepository.java
@@ -2,7 +2,7 @@ package kr.di.uoa.gr.jedaiwebapp.models;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface WorkflowResultsRepository extends CrudRepository<WorkflowResults, Double>{
+public interface WorkflowResultsRepository extends CrudRepository<WorkflowResults, Integer>{
 
 	WorkflowResults findById(int id);
